Correção do erro 500 ao retornar o TipoPrestador por ID...

+++ src/main/java/br/com/fatesg/fabrica/projetofabrica/controlador/TipoPrestadorResource.java
@@ -34,7 +34,7 @@ public class TipoPrestadorResource {
 	}
 	
 	@GetMapping("/{id}")
-	public ResponseEntity<TipoPrestador> buscar(@PathVariable int id) {
+	public ResponseEntity<TipoPrestador> buscar(@PathVariable long id) {
 		TipoPrestador obj = negocio.findById(id);		
 		if (obj == null) {
 			return ResponseEntity.notFound().build();
