+++ src/main/java/br/com/fatesg/fabrica/projetofabrica/repositorio/TipoPrestadorRepository.java
@@ -11,7 +11,7 @@ import br.com.fatesg.fabrica.projetofabrica.TipoPrestador;
 @Repository
 public interface TipoPrestadorRepository extends JpaRepository<TipoPrestador, Long> {
        
-	TipoPrestador findById(@Param("id") int id);
+	TipoPrestador findById(@Param("id") long id);
                
 	TipoPrestador findDescricaoById(@Param("id") Long id);
